+++ typescript/infra/scripts/funding/fund-keys-from-deployer.ts
@@ -91,6 +91,10 @@ const RC_FUNDING_DISCOUNT_DENOMINATOR = ethers.BigNumber.from(10);
 const CONTEXT_FUNDING_TIMEOUT_MS = 5 * 60 * 1000; // 5 minutes
 const CHAIN_FUNDING_TIMEOUT_MS = 1 * 60 * 1000; // 1 minute
 
+// Need to ensure we don't fund non-vanguard chains in the vanguard contexts
+const VANGUARD_CHAINS = ['base', 'arbitrum', 'optimism', 'ethereum', 'bsc'];
+const VANGUARD_CONTEXTS: Contexts[] = [Contexts.Vanguard0];
+
 // Funds key addresses for multiple contexts from the deployer key of the context
 // specified via the `--context` flag.
 // The --contexts-and-roles flag is used to specify the contexts and the key roles
@@ -259,6 +263,18 @@ class ContextFunder {
     roleKeysPerChain = objFilter(
       roleKeysPerChain,
       (chain, _roleKeys): _roleKeys is Record<Role, BaseAgentKey[]> => {
+        // Skip funding for vanguard contexts on non-vanguard chains
+        if (
+          VANGUARD_CONTEXTS.includes(this.context) &&
+          !VANGUARD_CHAINS.includes(chain)
+        ) {
+          logger.warn(
+            { chain, context: this.context },
+            'Skipping funding for vanguard context on non-vanguard chain',
+          );
+          return false;
+        }
+
         const valid =
           isEthereumProtocolChain(chain) &&
           multiProvider.tryGetChainName(chain) !== null;