+++ typescript/infra/test/warp-configs.test.ts
@@ -3,8 +3,8 @@ import chaiAsPromised from 'chai-as-promised';
 
 import { DEFAULT_GITHUB_REGISTRY } from '@hyperlane-xyz/registry';
 import { getRegistry } from '@hyperlane-xyz/registry/fs';
-import { MultiProvider } from '@hyperlane-xyz/sdk';
-import { diffObjMerge, rootLogger } from '@hyperlane-xyz/utils';
+import { HypTokenRouterConfig, MultiProvider } from '@hyperlane-xyz/sdk';
+import { rootLogger } from '@hyperlane-xyz/utils';
 
 import { getWarpConfig, warpConfigGetterMap } from '../config/warp.js';
 import {
@@ -47,24 +47,28 @@ describe('Warp Configs', async function () {
 
   for (const warpRouteId of warpIdsToCheck) {
     it(`should match Github Registry configs for ${warpRouteId}`, async function () {
-      const warpConfig = await getWarpConfig(
-        multiProvider,
-        envConfig,
-        warpRouteId,
-      );
-      const { mergedObject, isInvalid } = diffObjMerge(
-        warpConfig,
-        configsFromGithub![warpRouteId],
-      );
-
-      if (isInvalid) {
-        console.log('Differences', JSON.stringify(mergedObject, null, 2));
+      const warpConfig: Record<
+        string,
+        Partial<HypTokenRouterConfig>
+      > = await getWarpConfig(multiProvider, envConfig, warpRouteId);
+      for (const key in warpConfig) {
+        if (warpConfig[key].mailbox) {
+          delete warpConfig[key].mailbox;
+        }
+      }
+      const expectedConfig = configsFromGithub![warpRouteId];
+      for (const key in expectedConfig) {
+        if (expectedConfig[key].mailbox) {
+          delete expectedConfig[key].mailbox;
+        }
       }
 
-      expect(
-        isInvalid,
-        `Registry config does not match Getter for ${warpRouteId}`,
-      ).to.be.false;
+      expect(warpConfig).to.have.keys(Object.keys(expectedConfig));
+      for (const key in warpConfig) {
+        if (warpConfig[key]) {
+          expect(warpConfig[key]).to.deep.equal(expectedConfig[key]);
+        }
+      }
     });
   }