+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @hyperlane-xyz/cosmos-types
 
+## 12.2.0
+
 ## 11.0.0