+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hyperlane-xyz/infra
 
+## 12.2.0
+
+### Patch Changes
+
+- Updated dependencies [c7934f711]
+- Updated dependencies [ecbacbdf2]
+  - @hyperlane-xyz/sdk@12.2.0
+  - @hyperlane-xyz/helloworld@12.2.0
+  - @hyperlane-xyz/utils@12.2.0
+
 ## 12.1.0
 
 ### Minor Changes